Begun working on converting to a Stack<Object> for the VMEnvironment stack.
Soon we can work with Strings as variables....

+++ src/main/java/org/mangorage/mangolang/instruction/impl/arithmetic/Add.java
@@ -7,8 +7,8 @@ import org.mangorage.mangolang.vm.VMEnvironment;
 @AutoRegisterInstruction
 public final class Add implements Instruction {
     public void execute(VMEnvironment env) {
-        int b = env.getStack().pop();
-        int a = env.getStack().pop();
+        int b = (int) env.getStack().pop();
+        int a = (int) env.getStack().pop();
         env.getStack().push(a + b);
     }
 }
