+++ src/main/java/org/mangorage/mangolang/instruction/impl/comparison/Equals.java
@@ -8,8 +8,8 @@ import org.mangorage.mangolang.vm.VMEnvironment;
 public final class Equals implements Instruction {
     @Override
     public void execute(VMEnvironment env) {
-        int b = env.getStack().pop();
-        int a = env.getStack().pop();
+        int b = (int) env.getStack().pop();
+        int a = (int) env.getStack().pop();
         env.getStack().push(a == b ? 1 : 0);
     }
 }
